feat: Update eLife/api package to version 2.25.0 and add deprecation warnings
This commit updates the 'elife/api' package to version 2.25.0 in the
'composer.json' and 'composer.lock' files. Additionally, deprecation warnings
have been added to the 'src/bootstrap.php' and 'test/SmokeTest.php' files for
older versions of the application's APIs. A few code clean-ups are also
included in this commit.