Add reviewed-preprint in promotional collections (#246)
* Add a reviewed-preprint in promotional collections

* clean up

* v1 support added

* feat: Add deprecation warnings and version requirement checks

Add deprecation warnings to the API responses for collections and promotional collections that are using deprecated versions. Also, add checks to enforce the minimum required version for certain collections and promotional collections. Update the smoke tests to reflect these changes.

* feat: Add deprecation warnings for certain API versions

This commit includes deprecation warnings for certain versions of the API in the response headers. These warnings will be displayed when accessing blog articles, collections, labs posts, press packages, and promotional collections. Additionally, this commit includes tests to verify the correct functionality of these warnings.

* feat: Update eLife/api package to version 2.25.0 and add deprecation warnings

This commit updates the 'elife/api' package to version 2.25.0 in the
'composer.json' and 'composer.lock' files. Additionally, deprecation warnings
have been added to the 'src/bootstrap.php' and 'test/SmokeTest.php' files for
older versions of the application's APIs. A few code clean-ups are also
included in this commit.

---------

Co-authored-by: Nathan Lisgo <nathan@lisgo.co.uk>